src/requirements.txt
## The following requirements were added by pip freeze:
cycler==0.10.0
graphviz==0.11.1
kiwisolver==1.1.0
matplotlib==3.0.0
nltk==3.4.4
numpy==1.16.4
Pillow==6.1.0
pyparsing==2.4.0
python-dateutil==2.8.0
singledispatch==3.4.0.3
six==1.12.0
wrapt==1.11.2
